refactor: extract useProfileLaunch hook from MainLayout
Made-with: Cursor
